+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 7.1.4 (unreleased)
 ------------------
 
+- If configured with include_textline_fields, enable search and replace on string fields (Archetypes IStringField and dexterity TextLine) 
+  [gotcha]
+
 - Translate field names in preview table
   [gotcha]